--- backend/api/database/models/__init__.py
@@ -1,1 +1,0 @@
-from api import db as db
