--- backend/api/database/schemas/__init__.py
@@ -1,1 +1,0 @@
-from ....api import ma as ma
