+++ backend/api/models/cell.py
