+++ backend/api/models/sensor.py
