+++ backend/api/resources/cell_data.py
@@ -1,10 +1,10 @@
 from flask import request, jsonify
 from flask_restful import Resource
 import pandas as pd
-from ..database.schemas.get_cell_data_schema import GetCellDataSchema
-from ..database.models.power_data import PowerData
-from ..database.models.teros_data import TEROSData
-from ..database.models.sensor import Sensor
+from ..schemas.get_cell_data_schema import GetCellDataSchema
+from ..models.power_data import PowerData
+from ..models.teros_data import TEROSData
+from ..models.sensor import Sensor
 from functools import reduce
 from io import StringIO
 from celery import shared_task
