+++ backend/api/resources/cell_id.py
@@ -1,8 +1,8 @@
 from flask_restful import Resource
 
 # from ..conn import engine
-from ..database.models.cell import Cell
-from ..database.schemas.cell_schema import CellSchema
+from ..models.cell import Cell
+from ..schemas.cell_schema import CellSchema
 
 cells_schema = CellSchema(many=True)
