+++ backend/api/resources/users_data.py
@@ -1,7 +1,7 @@
 from flask_restful import Resource
-from ..database.models.user import User
-from ..database.schemas.user_data_schema import UserDataSchema
-from ..auth.auth import authenticate
+from ..models.user import User
+from ..resources import authenticate
+from ..schemas.user_data_schema import UserDataSchema
 
 user_schema = UserDataSchema(only=(["email", "first_name", "last_name"]))
