+++ backend/api/schemas/__init__.py
@@ -1,0 +1,1 @@
+from ...api import ma as ma
