+++ backend/api/schemas/cell_schema.py
