+++ backend/api/schemas/get_cell_data_schema.py
