+++ backend/api/schemas/power_data_schema.py
