--- backend/setup.py
@@ -1,3 +1,0 @@
-from setuptools import setup, find_packages
-
-setup(name="api", packages=find_packages())
